Update examples.rst
I think I found the issue with line 281 - should it be dca.PLYield instead of dca.Yield ?

I now, however, get an error on line 316 with the numerically evaluated cum gas. If I change q_g to be simply q in this line, the code continues, but (a) I'm not sure this is what was intended and (b) I then get a subsequent error with teh quad integration.